# Rework of MachInst isel, branch fixups and lowering, and block ordering.