--- a/cranelift/codegen/src/isa/mod.rs
+++ b/cranelift/codegen/src/isa/mod.rs
@@ -77,8 +77,8 @@ mod riscv;
 #[cfg(feature = "x86")]
 mod x86;
 
-//#[cfg(feature = "x64")]
-//mod x64;
+#[cfg(feature = "x64")]
+mod x64;
 
 #[cfg(feature = "arm32")]
 mod arm32;