… cranelift/codegen/src/isa/aarch64/inst/mod.rs
… cranelift/codegen/src/isa/aarch64/lower.rs
… cranelift/codegen/src/isa/aarch64/lower_inst.rs
… cranelift/codegen/src/isa/aarch64/mod.rs
… cranelift/codegen/src/isa/x64/inst/args.rs
… cranelift/codegen/src/isa/x64/inst/emit.rs
… cranelift/codegen/src/isa/x64/inst/emit_tests.rs
… cranelift/codegen/src/isa/x64/inst/mod.rs
… cranelift/codegen/src/isa/x64/lower.rs
… cranelift/codegen/src/isa/x86/mod.rs
… cranelift/codegen/src/machinst/blockorder.rs
… cranelift/codegen/src/machinst/buffer.rs
… cranelift/codegen/src/machinst/compile.rs
… cranelift/codegen/src/machinst/lower.rs
… cranelift/codegen/src/machinst/mod.rs
… cranelift/codegen/src/machinst/sections.rs
… cranelift/codegen/src/machinst/vcode.rs
… cranelift/codegen/src/num_uses.rs
… cranelift/filetests/filetests/vcode/aarch64/arithmetic.clif
… cranelift/filetests/filetests/vcode/aarch64/condbr.clif
… cranelift/filetests/filetests/vcode/aarch64/jumptable.clif
… cranelift/filetests/filetests/vcode/aarch64/saturating-ops.clif
… cranelift/filetests/filetests/vcode/aarch64/shift-rotate.clif
… crates/environ/src/cranelift.rs