@bjorn3 2020-05-18
> I see debug! calls in potentially hottish places, eg MachBuffer::put1(). Are we sure those become zero cost in release builds?

`trace!` is probably more appropriate.