@julian-seward1 2020-05-19
One other comment (not to block landing):

> >  (revised isel driver): again, this is now more complex than it was. Question: is lookback past block boundaries still allowed? [..]

> Yes, that's right; [..]

FTR, I am still of the opinion that lookback outside the block, even for just pure values, will turn out to be a net loss in the end, in the sense that the extra register pressure will cause much more of a loss than any minor improvements in insn selection that might result. That said, I don't have any Actual Evidence to substantiate my claim, at least currently.